New script `find_cotranslational_folders.py`
This script identifies proteins that are likely to fold cotranslationally on the basis of native structure energy analysis and rare codon enrichment. The script uses `data.csv` and prints the names of genes where evolutionarily conserved rare codons occur 20-60 residue downstream of drops in the native state free energy along the gene transcripts.